DPR2-895: Use micronaut to enable or disable operational datastore functionality.
