+++ src/it/java/uk/gov/justice/digital/job/batchprocessing/BatchProcessorIT.java
@@ -21,7 +21,7 @@ import uk.gov.justice.digital.service.ValidationService;
 import uk.gov.justice.digital.service.ViolationService;
 import uk.gov.justice.digital.service.operationaldatastore.OperationalDataStoreConnectionDetailsService;
 import uk.gov.justice.digital.service.operationaldatastore.OperationalDataStoreDataAccess;
-import uk.gov.justice.digital.service.operationaldatastore.OperationalDataStoreService;
+import uk.gov.justice.digital.service.operationaldatastore.OperationalDataStoreServiceImpl;
 import uk.gov.justice.digital.service.operationaldatastore.OperationalDataStoreTransformation;
 import uk.gov.justice.digital.test.BaseMinimalDataIntegrationTest;
 import uk.gov.justice.digital.test.InMemoryOperationalDataStore;
@@ -81,7 +81,6 @@ class BatchProcessorIT extends BaseMinimalDataIntegrationTest {
 
     @BeforeEach
     public void setUp() throws Exception {
-        givenOperationalDataStoreWriteIsEnabled();
         givenDatastoreCredentials();
         givenPathsAreConfigured();
         givenRetrySettingsAreConfigured(arguments);
@@ -253,8 +252,8 @@ class BatchProcessorIT extends BaseMinimalDataIntegrationTest {
         CuratedZoneLoad curatedZoneLoad = new CuratedZoneLoad(arguments, storageService, violationService);
         OperationalDataStoreTransformation operationalDataStoreTransformation = new OperationalDataStoreTransformation();
         OperationalDataStoreDataAccess operationalDataStoreDataAccess = new OperationalDataStoreDataAccess(connectionDetailsService);
-        OperationalDataStoreService operationalDataStoreService =
-                new OperationalDataStoreService(arguments, operationalDataStoreTransformation, operationalDataStoreDataAccess);
+        OperationalDataStoreServiceImpl operationalDataStoreService =
+                new OperationalDataStoreServiceImpl(operationalDataStoreTransformation, operationalDataStoreDataAccess);
         underTest = new BatchProcessor(structuredZoneLoad, curatedZoneLoad, validationService, operationalDataStoreService);
     }
 
@@ -288,10 +287,6 @@ class BatchProcessorIT extends BaseMinimalDataIntegrationTest {
                         credentials
                 )
         );
-    }
-
-    private void givenOperationalDataStoreWriteIsEnabled() {
-        when(arguments.isOperationalDataStoreWriteEnabled()).thenReturn(true);
     }
 
     private void thenStructuredCuratedAndOperationalDataStoreContainForPK(String data, int primaryKey) throws SQLException {
